requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0

# Database (sqlite3 is built into Python)

# Data processing
scikit-learn>=1.3.0

# Visualization
streamlit>=1.28.0
plotly>=5.17.0

# AI/LLM
openai>=1.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: Async support
aiohttp>=3.8.0

# Optional: Additional data sources
ccxt>=4.0.0
